[heroku-client] removed default exports
