+++ types/heroku-client/heroku-client-tests.ts
@@ -1,4 +1,4 @@
-import Heroku from 'heroku-client';
+import Heroku = require('heroku-client');
 
 const heroku = new Heroku({ token: process.env.HEROKU_API_TOKEN });
